Port over and substantially modify the U-Boot AHCI driver.
Remove the glue code that made AHCI drives pretend to be SCSI drives, refactor
the code to structure it a bit better, and adapt it to use the facilities
available from libpayload instead of the U-Boot libraries.

This code is currently hardcoded to keep track of a single AHCI device. That
will be changed in the future.

BUG=chrome-os-partner:8339
TEST=With this and other changes, built and booted on Stumpy and accessed the
SSD successfully.
BRANCH=None

Change-Id: I358238e25b97b1181c3608ba2209293614f8aec4
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/24872
Reviewed-by: Stefan Reinauer <reinauer@google.com>