Make vboot's disk related functions use the new AHCI driver.
The code is currently hardcoded to use the single AHCI drive set up by the
AHCI driver and nothing else. That will be changed in the future, but is
sufficient for now and unblocks getting other things working.

BUG=chrome-os-partner:8339
TEST=With this and other changes, built and booted on Stumpy and accessed
the SSD successfully. Vboot was able to verify one of the kernel partitions.
BRANCH=None

Change-Id: I1f9fab2a3335386d8a751344faadced290282e78
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/24885
Reviewed-by: Stefan Reinauer <reinauer@google.com>