Skip past empty SATA links using the DET field of the SSTATUS register.
Instead of waiting a full second for the DET field to indicate a present
device for every link, wait one millisecond first and then assume anything
that's not detected isn't going to be.

The DET field is based on the PHY layer of the SATA stack, and should be set
very quickly (on the order of several microseconds) from the time reset is
initiated. Waiting a full second for each port in turn is very, very
conservative and slows down boot considerably.

BUG=chrome-os-partner:8339
TEST=Built and booted on Stumpy.
BRANCH=None

Change-Id: I1eb9b79f988ff4d08adb627385f30e81dbbee320
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/24967
Reviewed-by: Stefan Reinauer <reinauer@google.com>