Parallelize waiting for SATA ports to spin up.
This change tells all the SATA ports to spin up, then waits 10 ms, and then
checks which ones seem to be alive. This may not be correct because I can't
find anything which sets a limit on how long it should take to establish
physical communication, but empirically on a small sample (one machine) this
seems to work.

BUG=chrome-os-partner:8339
TEST=Built and booted on Stumpy and Link.
BRANCH=None

Change-Id: I3a1f32979dbb845cf75b6a99f9e5d5f6c56d7826
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/24968
Reviewed-by: Stefan Reinauer <reinauer@google.com>