depthcharge: Wipe memory when requested.
This change imports the memory wiping code from U-Boot and hooks it up in
depthcharge's main function. The code which orchestrates the wiping is put
inline in vboot_init which is not ideal. Moving it elsewhere will require some
adjustments to some other things like include paths and should happen soon.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I077821492b38f6fc796930fd45a22aec96e8c20d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26147
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Duncan Laurie <dlaurie@google.com>