depthcharge: Add more subdirectories, and change how includes work.
This change adds some new subdirectories, makes includes relative to the base
of the source directory, splits out the "time.c" vboot file into the vboot
interface part and the time keeping part, and makes all the includes in C
files and the include files themselves follow some conventions. The "include"
directory is no longer used or needed and is deleted. Finally, a bunch of
unnecessary includes of libpayload.h and libpayload-config.h were removed.

The new subdirectories are:
base: Basic functionality which doesn't fit elsewhere.
boot: Files related to actually booting a Linux kernel.
drivers: Files related to controlling devices.

The conventions for including files is now to include all files not provided
by depthcharge first using <>s, followed by includes which are provided by
depthcharge using ""s, seperated by a blank line if applicable. Each group
should be locally alphabetized.

In the include files themselves, the compiler guard should start with two
underscores, then have the path of the include with slashes replaced with
underscores, then the name of the file with any dots replaced with underscores,
and finally two more underscores. For instance:

 base/foo.h => __BASE_FOO_H__

The #endif which ends the file should have the compiler guard after it in a
/**/ style comment, like this:

 #endif /* __BASE_FOO_H__ */

BUG=chrome-os-partner:8339
TEST=Built and booted on Link
BRANCH=None

Change-Id: I4187b22e644a5dc1ac5a76f6f885e6ed4e4186a6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26164
Reviewed-by: Stefan Reinauer <reinauer@google.com>