depthcharge: Rename the ambiguous firmware.c to load_firmware.c.
The original name of the file holding the VbExHashFirmwareBody function was
simply "firmware.c" which was pretty ambiguous. All of depthcharge is firmware,
so really anything could go in there. This changes it to "load_firmware.c"
which is a bit more specific, and the name used for a similar file in U-Boot.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I8aeeb08bef4d52042094b1ef1ebf11d788dd24cf
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26199