depthcharge: libpayload: Change "gpios" in depthcharge into flags.
This change moves from treating values like the dev switch and write protect
as GPIOs to treating them as more abstract flags. Some abstraction was added
so that most code will ask for the value of a flag, and then the flag code will
know to check a GPIO, ask the EC, look things up in the sysinfo structure, etc.
Also, some constants which were used to interpret the contents of the coreboot
tables were moved to the appropriate libpayload header file. The constant which
describes the maximum length of a GPIO name was renamed to have a CB_ prefix.
That makes it more obvious what sort of GPIO name it describes, and reduces the
change of a name collision. It also makes it more consistent with other names
in that header, although some other exceptions still exist.

In the future, the front end flag code may need to be further decoupled from
the back end which actually fetches the value. I've elected to put that off for
now since practically speaking there's only one way to get values right now.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ida4ccc8b203765a8b5063a62ad78974d931d259a
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26332