depthcharge: Grab U-Boot's pch GPIO driver.
The version in depthcharge is significantly stripped down because it doesn't
have to support any particular API at the moment. It drops code which keeps
track of whether two people are using the same GPIO (not likely at this stage)
and also whether the PCI device it expects the GPIOs to hang off of looks ok.
The later check might be nice to have if we need this driver to be flexible and
attempt to run when things look like they might work. Alternatively we could
force it to only work with PCI device IDs which are known good which would
probably be safer.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I520f4d27e8c8ec6375ac8598f73e2c8e60587e2d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26346