depthcharge: Implement bitmap support.
This change adds some data structures which help decode bitmaps, and some code
that actually draws the bitmaps into the coreboot framebuffer. That support
augments the corebootfb support in libpayload and, to avoid namespace issues,
uses the dc_corebootfb prefix for its functions.

Only palettized version 3 bitmaps are supported, and that limits the bits per
pixel to 8. 8 bits per pixel has been tested, but other color depths
(specifically 1 and 4) have support implemented.

The code which extracts the palette index for a pixel has support for pulling
out values larger than 8 bits, but that would appear to be unnecessary because
the palette isn't used with those color depths. I chose to leave that support
in so that the code that pulls pixels out of the bitmap can be pulled out and
used both to pull out indices and to pull out unencoded data.

BUG=chrome-os-partner:8339
TEST=With this and another change that hooks up VbExDisplayImage, built and
booted on Link, halted the CPU after all bitmaps were drawn, and inspected the
image on screen as compared to one drawn by U-Boot. Verified visually that they
were the same. Booted into ChromeOS.
BRANCH=None

Change-Id: I20565814dbc86ed741e6fbd9811b7e588f88a28b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26534
Reviewed-by: Bill Richardson <wfrichar@google.com>
Reviewed-by: Stefan Reinauer <reinauer@google.com>