depthcharge: Move some code around, and add keyboard initialization.
This change moves the code that looks up flags relevant to vboot out of
vboot_init and into main which serves two purposes. First, it keeps vboot_init
from getting overly complicated or bogged down in stuff unrelated to calling
into vboot. Second, it makes it easier to figure out whether we need keyboard
support when in the main function so we can keep all the bits of
initialization code there.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Built and booted on Link, pressed space bar,
and verified that the machine went into recovery mode.
BRANCH=None

Change-Id: I54a48fa4cc01c2a7e40e3e6581f4ce18081e7e9d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26751
Reviewed-by: Stefan Reinauer <reinauer@google.com>