depthcharge: Teach VbExKeyboardRead the keyboard version of the arrow keys.
The original version of VbExKeyboardRead was taken from U-Boot and written to
expect escape sequences for the arrow keys. The libpayload keyboard driver
returns constants defined in curses.h, though, and need to be interpreted
differently.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Used arrow keys over the serial terminal and on
the device's keyboard to change the locale. Before this change they worked
over the serial console but not the keyboard. After it worked on both and saw
that they changed the locale as expected.
BRANCH=None

Change-Id: I7c1bbdf5e3a7d465a91cf066f34e05716cd116bc
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/26828
Reviewed-by: Stefan Reinauer <reinauer@google.com>