depthcharge: Keep track of USB storage devices.
This change implements the usbdisk_create and usbdisk_remove functions which
maintain a linked list of BlockDev structures that describe detected USB keys.
That list can be processed by the vboot interface code to tell it about
available recovery media.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link, verified messages about adding or removing a
disk were printed when a USB key was inserted and removed.
BRANCH=None

Change-Id: I699f4a01e769126896bb585b6197467d039fe5cb
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27890
Reviewed-by: Stefan Reinauer <reinauer@google.com>