depthcharge: Initial version of the mkbp driver.
This driver was taken from U-Boot and modified somewhat minimally to get it to
build and initialize in depthcharge. References to device tree, I2C and SPI
were all removed, a function which checks for an interrupt from the EC was
left empty, and some file names were changed to remove redundancy since
they're in a directory specific to the ChromeOS EC.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Verified that the MKBP driver seemed to
initialize properly, didn't throw any errors, and reported a reasonable
looking EC version string.
BRANCH=None

Change-Id: Ib521923ea3d6d04d322d3d0c80d25623b7a0ad81
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27001
Reviewed-by: Stefan Reinauer <reinauer@google.com>