depthcharge: Simplify and streamline the mkbp LPC code a little.
There aren't any huge changes here, but this does make the LPC code a little
more streamlined and compact and easier to read. This change also took out
some printfs which printed the data going to/from the EC. This used to be
gated on a #define, but was changed into a pure printf when the change was
brought in. It might be readded in the future, but for now it was a bit too
verbose and inconvenient to keep in.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link, verified that the EC version string was still
read properly.
BRANCH=None

Change-Id: I22f5eeaa448321816885923b9770d0b0bb725f66
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27003
Reviewed-by: Stefan Reinauer <reinauer@google.com>