depthcharge: Add a device tree with separate sections for the EC in the FMAP.
Depthcharge doesn't use the device tree, so it needs all the information about
things are in the image to be spelled out in the FMAP. That's not currently
the case with the device tree currently used by U-Boot and borrowed for
depthcharge. This change adds a new device tree which is just the fmap portion
of the U-Boot one, plus a few small bits cros_bundle_firmware needs to work.
It explicitly includes sections for the EC binary and vblock.

This scheme is less flexible than the device tree based one because there's a
single FMAP which lives in the image, and it's in the RO area. We'll never be
able to change the amount of space given to the EC vs. the system firmware
once things are locked down. There could be multiple device trees, however,
some of which could be in RW and could be changed to adjust that boundary.
It's not clear how U-Boot would handle such an inconsistency since the RO
firmware may need to access the EC or system firmware blobs, but it would at
least be closer.

This change at least gets things working, though, and avoids having any impact
on U-Boot and cros_bundle_firmware as it exists today.

BUG=chrome-os-partner:8339
TEST=Built the device tree with dtc, installed it and ran it through
cros_bundle_firmware by directing the chromeos-bootimage ebuild to use it.
BRANCH=None

Change-Id: I71510fcf7d4531be60b0f2d3a7a628decbe26fb2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27005
Reviewed-by: Stefan Reinauer <reinauer@google.com>