depthcharge: Ground work to support booting into RW firmware.
This change makes the depthcharge Makefile build separate RW and RO images.
The images are the same except that the RO version builds in a small,
independent bit of code that will become a trampoline for starting the RW
version. Currently the only thing in it is libgcc, and the wrapper functions
which adapt its ABI to the non-standard one depthcharge is built with. It's
linked to be where the kernel will eventually be loaded because we'll be done
using it by the time we need that space for the kernel, and because there's
minimal risk of running into anything permanent there since Linux would sit
on top of it eventually.

Both RW and RO versions are linked using the same linker script. The two are
different because the trampoline code isn't included in the RW version, and so
those sections are empty. The trampoline areas draw from files which end in
trampoline.o which include the files added to the trampoline class of sources,
and the trampoline.o file they're all bound together in. trampoline.o is run
through objcopy so that its symbols are prefixed with the string .tramp. That
way, they won't get mixed with code/data from elsewhere, and so that there
isn't any accidental linking across the two areas. The "." character won't
show up in any valid C identifiers, so it should be difficult (but not
impossible) to reference from one to the other.

This change also takes advantage of a quirk of cros_bundle_firmware and how it
builds Coreboot based images. It first takes the "uboot" file name passed in,
and attempts to figure out what the ELF file the uboot ebuild created is
called. First, it tries to drop the assumed ".bin" extension, and if that file
doesn't exist, it tries to replace the ".bin" with ".elf" It doesn't check
whether any substitution was actually made, or whether the resulting file
exists. It then takes whatever name it came up with and installs it into CBFS
as the coreboot payload. Meanwhile, the raw name, which is assumed to be the
binary version of the firmware, is glued into the image in both the RW
areas.

I want depthcharge's RO and RW versions to be different, so I take advantage
of this behavior and name the RO version ".elf" and the compressed RW version
".bin". This is a bit hacky since cros_bundle_firmware is assuming there's only
one version of the firmware, but it gets the job done without changing the flow
for any existing firmware.

In the longer term, it would be great if cros_bundle_firmware didn't have
these built in assumptions and could be more flexible when putting together
an image.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Looked at the RO and RW ELF files (RW after
decompression) and verified they size of their various sections were
reasonable and made sense when compared with the layout before this change.
Verified that the RO ELF could be stripped and that the .debug sections would
be dropped appropriately, even for the trampoline sections. Used hexedit to
verify that the contents of the RW sections of the image were the RW version
of depthcharge.
BRANCH=None

Change-Id: Icc139915e9bfbd3904d09998d6100ce0b53468ef
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27665
Reviewed-by: Stefan Reinauer <reinauer@google.com>