depthcharge: Move cparams somewhere it can be shared between RW and RO.
This change sets up special sections in the linker script for the cparams data
structure, and the shared data blob. They are put in a fixed location so they
can be found by the RW firmware easily. They're removed from the ELF with
objcopy so loading the RW firmware doesn't clobber the value left there by the
RO firmware.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I17d84a2d47a17fc8347f120f7d741b19bfb17341
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27666
Reviewed-by: Stefan Reinauer <reinauer@google.com>