depthcharge: Move initializing some hardware to the point we know we need it.
If we end up booting an RW firmware, we'll likely want to redo some hardware
initialization which may have been done in the RO firmware. This change delays
setting up hardware that isn't needed yet until we commit to which version
we're using. That not only avoids the delay of setting things up twice, it
also leaves the RW firmware a cleaner slate and more opportunity to avoid
correct bugs in the RO firmware.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ib768b9c8dce6aad6636de9fb7ef1dfbef0d52202
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27669
Reviewed-by: Stefan Reinauer <reinauer@google.com>