depthcharge: Implement the RW trampoline and use it when needed.
This change implements the trampoline which goes from the RO to RW version
of the firmware. The code which runs in context of the main part of the RO
firmware is responsible for decompressing the RW image and for checking it for
any errors. Unlike the trampoline, it's linked with libpayload and can report
problems back to the user.

When it gets control, the first thing the trampoline does is switch over to
its own stack, out of the way of the one which is part of the main image. It
then very simply loops over the segments described in the ELF program headers
and puts the pieces where they go. Once the new firmware is in place, it jumps
to its entry point.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Built with VbSelectFirmware's choice forcefully
overridden to select RW firmware A, and saw that the RW firmware started and
booted ChromeOS. Verified that the pieces of the RW firmware were where they
were supposed to be in the image and that symbols were properly segregated
between the trampoline and the normal image.
BRANCH=None

Change-Id: I98e3b2beede8be36bc8a776e0e84a2d8aa73b886
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27670
Reviewed-by: Stefan Reinauer <reinauer@google.com>