depthcharge: Place the global offset table purposefully in the linker script.
The global offset table sections of the image were "orphaned", or in other
words not explicitly layed out in the output in the linker script, and the
linker places it by making a "best guess" at where it belongs. To avoid it
guessing something unfortunate, this change tells it where to put those
sections both for the trampoline area and for the main body of the firmware.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link through the RW firmware.
BRANCH=None

Change-Id: I531a19813fed9d75d6b7727a3cd4a1e73d120ea7
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27671
Reviewed-by: Stefan Reinauer <reinauer@google.com>