depthcharge: Tell vboot about any USB drives we've found.
This further implements vboot's disk info gathering function so that it
reports information about USB disks when requested.

BUG=chrome-os-partner:8339
TEST=Built and booted. Booted into recovery mode and saw that vboot attempted
to read from a USB key when one was inserted. The read failed for some reason
so booting from the recovery key was unsuccessful, but the plumbing which told
vboot about the disk in the first place seems to work.
BRANCH=None

Change-Id: Ife8e7439e2ee92e91470a2f3dd74aeea19eb0fa0
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/27891
Reviewed-by: Stefan Reinauer <reinauer@google.com>