depthcharge: Move files out of "base" and into more appropriate homes.
The "base" directory had started to fill up with stuff that didn't have
anywhere else to be in the current directory structure. This change creates
three directories, image, vboot/util, and vboot/callbacks, and moves files
into them. image is for files related to managing the firmware image itself,
including entering the RW image from the RO image. vboot/callbacks is for
files which implement the vboot callback API. vboot/util is for other vboot
stuff which isn't a callback.

BUG=chrome-os-partner:8339
TEST=Built and booted on Lumpy.
BRANCH=None

Change-Id: I19637890d47a31615e115b68ebf7d1bd2af0d27f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28071
Reviewed-by: Stefan Reinauer <reinauer@google.com>