depthcharge: Add support for updating the ACPI tables provided by coreboot.
This change adds some code which updates the ACPI tables provided by coreboot
and fills in vboot related information it didn't know at the time it was
running. This code peeks into the vboot internal data structures which it
shouldn't, but some information is simply not accessible otherwise. This will
have to be fixed on the vboot side before it can be done properly here.

BUG=chrome-os-partner:8339
TEST=Built and booted on Lumpy, ran crossystem, saw reasonable values returned.
BRANCH=None

Change-Id: I27a3d2cb9cc219d6d3478beba8debee2ef0ac717
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28078
Reviewed-by: Duncan Laurie <dlaurie@google.com>