depthcharge: Make depthcharge build the vboot library locally.
The vboot_reference-firmware ebuild has gone away, so we need to build the
library locally like U-Boot does. Like U-Boot we override vboot's default
CFLAGS to be the same as ours, excluding include path options.

The vboot_reference Makefile doesn't seem to have any provisions for
abbreviated output, so when that's being built there's a lot of console spew.
That would be nice to fix in the future.

BUG=chrome-os-partner:8339
TEST=Built and booted on Lumpy.
BRANCH=None

Change-Id: I446d1ce2d7cebbba6b42d1519623f009319a5601
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28144
Reviewed-by: Stefan Reinauer <reinauer@google.com>