depthcharge: Print the right firmware ID when displaying debug text for vboot.
When vboot asks for depthcharge to print some debug info on the screen, it was
attempting to do the same thing as U-Boot, ie to print the read only and
active firmware IDs after the requested text. Since it wasn't easy to tell
which firmware was active before, the active firmware ID just printed a FIXME
placeholder instead. Now that we have some of that info lying around from when
we update the ACPI tables, we can grab it from there.

It was hacky for us to look into vboot's private data to find out which
firmware we are, and it's also hacky for us to look at the ACPI tables for the
same info, although slightly less so since at least it's not private data. In
both cases, it would be best if vboot itself managed that sort of data and just
asked for the firmware IDs when it needed them. It would be able to put them in
the debug text itself without us having to augment it.

BUG=chrome-os-partner:8339
TEST=Built and booted on Lumpy. Pressed tab at the developer screen, and saw a
reasonable looking active firmware ID. Since all firmware IDs are the same, I
only verified that it was the same as the one printed for read only.
BRANCH=None

Change-Id: I07429f4b03f30404127af1340d904b4d6f40cddb
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28155