depthcharge: Move the BINF_* constants into acpi.h.
These constants are part of the "public" interface to the ACPI data, unlike
the VDAT_* constants which describe some of the private vdat data. They're
needed to make sense of the ACPI data later, to determine whether we're in
read only, read/write a, or read/write b.

BUG=chrome-os-partner:8339
TEST=Built and booted on Lumpy.
BRANCH=None

Change-Id: Iaa400bcb5ef00a335576f5fd8d234220212c3801
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28154