depthcharge: Plumb the location and size of the kernel area to C.
This area had been independently set up in the linker script and in C, with no
representation of the areas size in the linker script. This change puts symbols
at the beginning and end of the kernel buffer which are named consistently with
other symbols which are already present, and adds them to the symbols.h header
file.

BUG=chrome-os-partner:8339
TEST=Built and booted on Lumpy.
BRANCH=None

Change-Id: I06710b1d4a7fd19bbc3fe2b8f40456394ef77e57
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28220
Reviewed-by: Duncan Laurie <dlaurie@google.com>