depthcharge: Use the biosincludes.h provided by vboot_reference.
Depthcharge was using its own biosincludes.h which was modeled off the one
U-Boot uses and which had only a single #define in it. This file is included
by sysincludes.h which is included by vboot_api.h, and, depending on what else
you've included by that point, that may be all you need.

When being used by vboot_reference, there's a header file called biosincludes.h
for each arch which has a bunch of other definitions in it. Because of the way
include paths are processed, when vboot_api.h is used by the host firmware,
its local version of biosincludes.h is used instead of the one vboot comes
with. This can cause problems if the single definition that was available
wasn't enough.

Recently, the vboot_reference-firmware ebuild was removed, and the vboot_fw.a
library started being built as part of the host firmware's build process,
depthcharge in this case. If the depthcharge CFLAGS value was used directly,
the compile would fail because certain definitions weren't available. If,
however, the -nostdinc compiler option and the include paths were filtered
out, the build would succeed. It's assumed that's because the system version
of some headers were being picked up, and they indirectly provided the missing
definitions.

The vboot_reference ebuild was changed very recently to install its versions
of biosincludes.h for all of its architectures into the target file system.
Rather than providing our own incomplete version of that header, or even
filling it out and attempting to maintain it in parallel to vboot's, this
change removes our version entirely, passes our CFLAGS to vboot unfiltered,
and adds in an include path that will pick up the biosincludes.h file from the
vboot_reference ebuild.

It also removes an extra include of stdint.h which was added so that the
subsequent include of gbb_header.h would find the standard integer types it
needed. Now that biosincludes.h has what it's supposed to in it, that's no
longer necessary.

One possible problem with this scheme is that it could be possible for the
type definitions in vboot's biosincludes.h to conflict with ones depthcharge
gets through stdint.h which could cause build failures. A possible solution
would be for biosincludes.h to include stdint.h instead of making its own
types, but that presumes the existence of stdint.h. Another possible solution
would be for biosincludes.h to define its own types with slightly different
names like, perhaps, vb_uint8_t. Defining its own types and expecting them to
be certain sizes is dangerous regardless since I think the sizes of built in
integer types varies and is up to the implementators of the compiler to some
extent.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ia08f16244ee0298db8c16bf9059cbb4de8a37275
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28315
Reviewed-by: Duncan Laurie <dlaurie@google.com>