depthcharge: Add a call to usb_poll before returning info about USB disks.
Because vboot checks for keypresses frequently and that causes a call to
usb_poll, depthcharge would discover new USB devices fairly quickly and have
their info to return to vboot when requested. This change adds an explicit
call to usb_poll to remove that unintentional dependency.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Rebooted into recovery mode and checked that a
USB key was detected.
BRANCH=None

Change-Id: I08f498fbce39e064afe4b943e032564c88b1ff25
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28519
Reviewed-by: Stefan Reinauer <reinauer@google.com>