depthcharge: Add some generic support for doubly linked lists.
These are somewhat inspired by the way Linux manages lists like this where the
actual list nodes are embedded in other data structures and a container_of
macro is used to generate a pointer to the parent. container_of is a fairly
generic and useful macro, so at some point it would make sense to move it to a
more generic place.

BUG=chrome-os-partner:8339
TEST=Built and booted with this change and another that uses this new code.
BRANCH=None

Change-Id: I43f8065a94d2bf8398e27da1a6beeee0309600b2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28522
Reviewed-by: Stefan Reinauer <reinauer@google.com>