depthcharge: Use a doubly linked list to keep track of block devices.
This makes it easier for the USB code to remove devices from the list. It also
makes it more feasible to keep track of all devices of a type, for instance
removeable and non-removeable devices, on centralized lists which aren't owned
by the USB, AHCI, etc., drivers.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Did a recovery mode boot and inserted a USB key.
Saw consistent behavior before and after the boot. With extra changes which fix
the mass storage controller support, booted off of USB.
BRANCH=None

Change-Id: Ie3d33f074ef91100a6f06f5ab75e2b8793d2643e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28524