depthcharge: Fix touch devices.
Strings associated with a DMI table, if any, are stored after the formatted
portion of the table. The table as a whole, including the strings, is
terminated with two zero bytes. The DMI tables coreboot provides leave one of
the strings empty to be filled in later by U-Boot/depthcharge, and fill it
with space characters. That way a portion of the spaces can be overwritten
with a NULL terminated string, and the spaces will pad out the structure to
its full lenght without adding a double NULL which would end it prematurely.

The code in depthcharge which filled in this field wasn't written with this
space padding in mind, and simply memcpy-ed the full size of the field in the
image, as defined by the FMAP, into where the vbt10 pointer was pointing. This
produced a valid NULL terminated string, but it also clobbered all of the space
padding and introduced a double NULL before the end of the table. The third and
subsequent NULLs were interpreted as the start of the next table, which is
invalid. This made the DMI tables as a whole invalid.

When deciding which platform devices to enable during boot, the kernel would
look at the DMI tables to determine what it was running on and what
unprobe-able devices (I2C) it needed to set up. Because the DMI tables were
effectively invalid, it errored out, and didn't install the drivers for the
touch devices.

This change replaces the memcpy with something that stops copying after the
NULL terminator to preserve the spaces. It's written to avoid any unbounded
string operations like strcpy, and to ensure that there is always an
appropriately placed NULL at the end of the firmware ID.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Without this change, the touchpad didn't work.
With this change, the touchpad works.
BRANCH=None

Change-Id: I47d9b0b4178d1e078a40b47e12a593f461691c26
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28525