depthcharge: Reboot when vboot returns anything other than success.
The comments in vboot_api.h say that the caller should reboot if vboot returns
anything other than success for a number of functions. Depthcharge was
selectively rebooting when it saw certain return values and returning an error
internally which would cause a halt.

BUG=chrome-os-partner:8339
TEST=Built and booted on a Link where the TPM had incorrectly advanced firmware
version numbers. Before this change the firmware halted, and after this change
the system rebooted into recovery mode correctly.
BRANCH=None

Change-Id: I07d144192e5f257d808efff303bf9fa673ab2727
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28740
Reviewed-by: Stefan Reinauer <reinauer@google.com>