depthcharge: Fix and reorganize the ICH GPIO driver.
The ICH GPIO driver, adapted from the one in U-Boot, had never been thoroughly
tested. Recent testing exposed a number of bugs in the driver which are fixed
by this change. I also took the opportunity to reorganize the driver a little
bit to try to consolidate the code which breaks a GPIO number down into bank
and bit numbers.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Saw that Link now recognizes that dev mode was
requested. It fails to recognize the user pressing enter to confirm which is
probably an unrelated bug.
BRANCH=None

Change-Id: Icdbc58a801db032c8c6c8fb22cc5daf2855ded41
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28745
Reviewed-by: Stefan Reinauer <reinauer@google.com>