depthcharge: Initialize console input devices on demand.
This change adds a fake console input driver who's sole purpose is to run
the initialization functions of other drivers when something is requesting
characters of input. That will avoid unnecessary initialization of hardware
without having to guess when, for instance, vboot is going to need the
keyboard.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Went into dev mode and pressed Ctrl+D and saw
that it was recognized.
BRANCH=None

Change-Id: Ia650df2bfb2bcbe538d2ca61e94cbf985b0a1dcf
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28967