depthcharge: Move reading of flags back into vboot_init.
The flags had moved out of vboot_init and into main in large part because they
were used to infer information about the state of vboot and what it was going
to try to do. Now that that's determined in other ways, the code can be moved
back to where it started. It's better to have it in vboot_init because then
changes to what and how information is gather for vboot initialization can be
made in one place.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ie405e7c4687378bb46817cdde0b1ffebeb596214
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28968