depthcharge: Build the code to start RW firmware into RO and RW images.
This is to prepare to move the code that starts an RW firmware out of main and
into vboot_select_firmware. vboot_select_firmware will only be called from the
RO code, and the linker should be smart enough to throw away that, and
start_rw_firmware which will only be called from there, in the RW image.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I4020e7f92a8e2eea9333fce0c33589e137c81d38
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28970