depthcharge: Move mkbp_init above acpi_update_data.
acpi_update_data tries to determine if it can trust the EC, and in the process
checks if there's an MKBP device pointer set. mkbp_init sets that pointer, so
it needs to run first. When they're in the other order, the acpi_update_data
errors out because of the failed call and boot stops.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Forced boot into RW firmware A. Without this
change, saw the boot halt after printing some error messages. With this change,
saw ChromeOS boot successfully.
BRANCH=None

Change-Id: I992187f8de226cdf774c5b4f220ac0ba65e77d43
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28959
Reviewed-by: Ronald G. Minnich <rminnich@google.com>
Reviewed-by: Stefan Reinauer <reinauer@google.com>