depthcharge: Use vboot internal data to determine the main firmware type.
It's apparently not sufficient to check if the boot time developer boot switch
is set in the coreboot tables to determine if we're in developer mode or not.
That was always a bit of an approximation anyway, and it turns out it's not a
very good one. U-Boot uses a flag returned by VbInit to figure that out, but
we aren't running that in the RW firmware, so we'd have to keep track of it
across firmwares, etc. Since we're already using vboot internal structures to
fill in some crossystem data in the ACPI tables, this change simply extends
that to also figure out the main firmware type. Since we're no longer trying
to determine that ourselves, the only parameter to acpi_update_data, the
firmware type, is removed.

It would be possible to move the code that jumps to an RW firmware out of
ro_main and put it in vboot_select_firmware in stages.c and to remove the one
parameter from that function as well, a pointer to return the firmware
selected. Unfortunately that would force the linker to include all the code to
jump to an RW firmware in the RW firmware itself which would waste space since
in reality we'd never do that.

BUG=chrome-os-partner:8339
TEST=Before this change, crossystem would claim the main firmware type was
normal when in developer mode. After this change, it correctly reports the type
in developer mode, normal mode, and recovery mode.
BRANCH=None

Change-Id: Ie42497e9826d6e1005bb1b7caca56ea84802f527
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28964
Reviewed-by: Ronald G. Minnich <rminnich@google.com>
Reviewed-by: Stefan Reinauer <reinauer@google.com>